+++ src/ledc.rs
@@ -384,6 +384,15 @@ impl<'d> LedcDriver<'d> {
         }
         Ok(())
     }
+
+    #[cfg(not(any(esp32, esp_idf_version_major = "4")))]
+    /// Stop LED fading before target duty cycle is reached
+    pub fn fade_stop(&mut self) -> Result<(), EspError> {
+        unsafe {
+            esp!(ledc_fade_stop(self.speed_mode, self.channel()))?;
+        }
+        Ok(())
+    }
 }
 
 impl Drop for LedcDriver<'_> {